requirements.txt
# Follow The Goat - Dependencies
# ==============================

# Database
duckdb>=1.0.0
psycopg2-binary>=2.9.9  # PostgreSQL for archive database
pymysql>=1.1.0  # For one-time MySQL migrations

# Scheduling (NO .bat files - use this instead)
APScheduler>=3.10.0

# HTTP requests
requests>=2.31.0
httpx>=0.27.0

# Web API (FastAPI for data engine and webhook)
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Legacy Flask API (still used by some features)
Flask>=3.0.0
flask-cors>=4.0.0

# WebSocket (Binance order book stream)
websocket-client>=1.7.0

# Data processing
pandas>=2.0.0
pyarrow>=14.0.0  # Fast DuckDB insertion via zero-copy Arrow tables

# Environment management
python-dotenv>=1.0.0

# Solana SDK (for swap testing)
solders>=0.18.0  # Solana Python SDK
base58>=2.1.1  # Base58 encoding for keys

